refactor: Create InputWrapper component
