+++ packages/ui/src/components/Input/Input/Input.tsx
@@ -1,41 +1,9 @@
 import React, { ElementRef } from 'react'
 import { useUuid } from '@aviato/hooks'
 import { ComponentProps } from '@stitches/react'
-import { styled } from '~/theme'
-import { Conditional } from '~/components/Utilities'
 
 import { BaseInput, BaseInputProps, StyledInput } from './BaseInput'
-import { IconError } from '~/icons'
-
-const InputWrapper = styled('div', {
-  width: '100%',
-})
-
-const Label = styled('label', {
-  fontSize: 15,
-  lineHeight: '32px',
-})
-
-const ErrorWrapper = styled('div', {
-  display: 'flex',
-  flexDirection: 'row',
-  marginTop: 8,
-})
-
-const IconWrapper = styled('div', {
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  height: '24px',
-  color: '$ErrorMain',
-  marginRight: 8,
-})
-
-const Error = styled('div', {
-  fontSize: 15,
-  lineHeight: '24px',
-  color: '$ErrorMain',
-})
+import { InputWrapper } from './InputWrapper'
 
 export interface InputProps extends BaseInputProps {
   label?: string
@@ -56,27 +24,13 @@ export const Input = React.forwardRef<
   const isInvalid = Boolean(error || invalid)
 
   return (
-    <InputWrapper>
-      <Conditional test={label}>
-        <Label>{label}</Label>
-      </Conditional>
-
+    <InputWrapper label={label} error={error}>
       <BaseInput
         ref={forwardedRef}
         id={uuid}
         invalid={isInvalid}
         {...remainingProps}
       />
-
-      <Conditional test={error}>
-        <ErrorWrapper>
-          <IconWrapper>
-            <IconError />
-          </IconWrapper>
-
-          <Error>{error}</Error>
-        </ErrorWrapper>
-      </Conditional>
     </InputWrapper>
   )
 })
